% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connection}
\usage{
connect(drv = RPostgres::Postgres(), dbname = "baxdb",
  host = "localhost", port = 5434, user = "postgres",
  password = "password")
}
\arguments{
\item{drv}{DBI driver}

\item{dbname}{String name of the database (default: gwasdb)}

\item{host}{Host of database (default: localhost)}

\item{port}{Number (default: 5434)}

\item{user}{String (default: postgres)}

\item{password}{String (default: password)}
}
\value{
A connection object
}
\description{
This is an example function that connects to
a local Docker instance of the GWAS database.
}
