% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect.R
\name{disconnect}
\alias{disconnect}
\title{Disconnection}
\usage{
disconnect(conn)
}
\arguments{
\item{conn}{Connection object}
}
\description{
This is an example function that connects to
a local Docker instance of the GWAS database.
}
