% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_bp_by_liketrait_for_species.R
\name{list_bp_by_liketrait_for_species}
\alias{list_bp_by_liketrait_for_species}
\title{List available traits}
\usage{
list_bp_by_liketrait_for_species(trait, species)
}
\arguments{
\item{trait}{String name of the trait of interest}

\item{species}{String name of the species of interest}
}
\value{
Dataframe results of query
}
\description{
List available traits
}
