% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_bp_by_liketrait_for_species.R
\name{list_bp_by_trait_for_species}
\alias{list_bp_by_trait_for_species}
\title{List the chromosome and SNP position for a given trait and species}
\usage{
list_bp_by_trait_for_species(trait, species, cutoff)
}
\arguments{
\item{trait}{String name of the trait of interest. This is an exact match.}

\item{species}{String name of the species of interest}

\item{cutoff}{Numeric (float) maximum value for model added p-value. The results will have a p-value less than or equal to the cutoff value specified.}
}
\value{
Dataframe results of query
}
\description{
Given a trait and species, return the chromosome and SNP position.
The trait string must be an exact match.
}
\seealso{
\code{\link{list_bp_by_liketrait_for_species}}
}
