% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_bp_by_liketrait_for_species.R
\name{list_bp_for_species}
\alias{list_bp_for_species}
\title{List the chromosome and SNP position for a given a species for all its traits}
\usage{
list_bp_for_species(species, cutoff = 0.05, model = "MaxCof")
}
\arguments{
\item{species}{String name of the species of interest}

\item{cutoff}{Numeric (float) maximum value for model added p-value. The results will have a p-value less than or equal to the cutoff value specified.}
}
\value{
Dataframe results of query
}
\description{
Given a species, return the chromosome and SNP position for all of its traits.
}
\seealso{
\code{\link{list_bp_for_species}}
}
