% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_traits_for_species}
\alias{list_traits_for_species}
\title{List the trait names for a given species}
\usage{
list_traits_for_species(species)
}
\arguments{
\item{species}{String binomial name of the species of interest}
}
\value{
Dataframe results of query
}
\description{
Given a species (binomial name), return the known traits
The species string must be an exact match.
}
