% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAllKnownTraits.R
\name{listAllKnownTraits}
\alias{listAllKnownTraits}
\title{List available traits}
\usage{
listAllKnownTraits(conn)
}
\arguments{
\item{conn}{Connection Object}
}
\value{
List
}
\description{
List all the known unique traits for all datasets
}
