% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listBasepairsByTraitForSpecies.R
\name{listBasepairsByTraitForSpecies}
\alias{listBasepairsByTraitForSpecies}
\title{List available traits}
\usage{
listBasepairsByTraitForSpecies(conn, params = list("Zea Maize"))
}
\arguments{
\item{conn}{Connection object to database}

\item{params}{Tuple containing species and trait to search by}
}
\value{
Dataframe results of query
}
\description{
List available traits
}
