% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query}
\usage{
query(conn, sql)
}
\arguments{
\item{conn}{Connection object to database}

\item{sql}{Custom query string}
}
\value{
Dataframe results of query as dataframe
}
\description{
This is a function intended to processing custom queries to the GWAS database.
}
