% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_bp_by_liketrait_for_species.R
\name{list_bp_by_liketrait_for_species}
\alias{list_bp_by_liketrait_for_species}
\title{List the chromosome and SNP position for a given trait and species}
\usage{
list_bp_by_liketrait_for_species(trait, species)
}
\arguments{
\item{trait}{String name of the trait of interest. This will match any trait that contains this value.}

\item{species}{String name of the species of interest}
}
\value{
Dataframe results of query
}
\description{
Given a trait and species, return the chromosome and SNP position
}
